Move escape handler to Modal component
